+++ plugins/CloudFile/src/Services/google_service.cpp
@@ -263,7 +263,7 @@ UINT CGDriveService::Upload(FileTransferParam *ftp)
 				CreateUploadSession(uploadUri, folderId, fileName);
 
 				uint64_t offset = 0;
-				size_t chunkCount = ceil(fileSize / chunkSize);
+				double chunkCount = ceil(double(fileSize) / chunkSize);
 				while (chunkCount--)
 				{
 					ftp->CheckCurrentFile();
